# server/go/model_envoy_fleet_item.go
/*
 * Kusk Gateway API
 *
 * This is the Kusk Gateway Management API
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type EnvoyFleetItem struct {

	Name string `json:"name"`

	Id string `json:"id"`

	Namespace string `json:"namespace"`
}

// AssertEnvoyFleetItemRequired checks if the required fields are not zero-ed
func AssertEnvoyFleetItemRequired(obj EnvoyFleetItem) error {
	elements := map[string]interface{}{
		"name": obj.Name,
		"id": obj.Id,
		"namespace": obj.Namespace,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertRecurseEnvoyFleetItemRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of EnvoyFleetItem (e.g. [][]EnvoyFleetItem), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseEnvoyFleetItemRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aEnvoyFleetItem, ok := obj.(EnvoyFleetItem)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertEnvoyFleetItemRequired(aEnvoyFleetItem)
	})
}